ncftp: fix host/target confusion
The ncftp build process tries to build and run a small program called
ccdv to beautify the build process output. If it manages to build and
run it, then it uses it.

Unfortunately, this doesn't work well when the target architecture is
close to the host architecture, but not exactly the same. Because both
architectures are close to each other, the test run of ccdv succeeds,
but real use of ccdv during ncftp build process causes an Illegal
instruction issue.

This for example happens with the CodeSourcery AMD64 toolchain, on a
build machine running an i7-4600U, and has been detected in the
autobuilders since the CodeSourcery AMD64 toolchain was upgraded at
the end of January:

  http://autobuild.buildroot.net/?reason=ncftp-3.2.6

The issue was also reported by Christopher Arguin back in July 2016:

  http://lists.busybox.net/pipermail/buildroot/2016-July/168026.html

and at the time, we identified that simply disabling the ccdv tool, by
passing --disable-ccdv, was enough to solve the issue. But Christopher
never submitted the patch, so the problem remained unfixed.

Therefore, we pass --disable-ccdv to the configure script, which
fixes:

  http://autobuild.buildroot.net/results/6eadad0e879ca70bb07b13b4196d42c64b11699f/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Reviewed-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
(cherry picked from commit 9647c3c5e9976c8ee666fdd91039289d09e07a4b)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>