mpd: fix build for ARC
mpd package fails for both internal and external ARC toolchain as check
for pthread support fails. Such checks fails because _REENTRANT flag is
not defined in gcc even when -pthread is passed.

So we add patch to gcc that defines _REENTRANT  on ARC when -pthread is
passed.

Also it disables mpd package for external ARC toolchain as it fails due
to the same issue.

This patch should be reverted as soon as the patch for GCC becomes a
part of ARC toolchain.

Fixes:
  http://autobuild.buildroot.net/results/7d7/7d70b62ad996830fbeca46dffcc7a1dc030e575d//

Signed-off-by: Vlad Zakharov <vzakhar@synopsys.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
(cherry picked from commit 4d3556662deba0014d9368a84344a59b5e167fac)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>