libiio: explicitly disable matlab bindings
If building on a machine with MATLAB installed, the MATLAB bindings will
default to ON, which is not the desired behavior for a cross build.

The bindings are designed to be called from within MATLAB, and we are
not currently running MATLAB on buildroot-generated targets.

This does not preclude the use of the bindings from a host connecting
over the network backend (assuming libiio on the host has the bindings
enabled).

Signed-off-by: Matthew Fornero <mfornero@mathworks.com>
Acked-By: Paul Cercueil <paul.cercueil@analog.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
(cherry picked from commit f358078b3374f3b316551660e89f5c463cb61608)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>