xlib_libXv: correct license info
The license is actually closer to ISC than MIT, E.G.:

https://opensource.org/licenses/ISC
vs
https://opensource.org/licenses/MIT

>From COPYING:
Permission to use, copy, modify, and distribute this software and its
documentation for any purpose and without fee is hereby granted,
provided that the above copyright notice appear in all copies and that
both that copyright notice and this permission notice appear in
supporting documentation, and that the names of Digital or MIT not be
used in advertising or publicity pertaining to distribution of the
software without specific, written prior permission.

E.G.  this is very similar to the normal ISC text which has been extended
with a BSD-3c style advertisinc clause.  Both are permissive licenses, but
it is more correct to call it ISC-like.

Notice: As pointed out by Rahul Bedarkar, this may actually be more closely
related to the OpenBSD template license:

http://cvsweb.openbsd.org/cgi-bin/cvsweb/src/share/misc/license.template?rev=HEAD

But that is also based on the ISC license, so calling it ISC-like is still
correct.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
(cherry picked from commit aad15918520391ea0952d742db9c52c3ac2b1f7e)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>