libnspr: fix Microblaze patch
Due to a mistake, the Microblaze patch was adding another #elif
defined(nios2), which doesn't make any sense. This commit gets rid of
it.

The rest of the noise in the change is due to the use of quilt to
generate the patches.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
(cherry picked from commit 7039c4d456259ca2c2708be16ebea98a82adee3f)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>