gst1-plugins-bad: fix webrtc option
When the WebRTC plugin option was introduced in commit
ee267886bc090335dd94c97dca1c294addccc090 ("gst1-plugins-bad: enable
webrtc plugin"), it was incorrect added as "webrtc", while the actual
name of the plugin and corresponding configure option is "webrtcdsp".

This commit therefore fixes the .mk file to use the correct name. And
also, since we want to keep Buildroot option consistent with the name
of the GStreamer plugins, it renames the Config.in option as well, and
introduces the necessary Config.in.legacy handling.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
(cherry picked from commit 4c06d2490a07f0b88f42c56c7409899fd2f5608a)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>