usb_modeswitch: explicitly set --host / --build for the embedded jimtcl
Fixes:
http://autobuild.buildroot.net/results/106/106af20196089e74c35daabdae46f030981600bd/

Otherwise the (handwritten) configure script uses a config.guess script from
2010 to figure out the build host, breaking builds on ppc64le hosts.

The result of --host / --build is only used to detect if we are building for
Windows or not, which is why things are working on x86(-64) build hosts
without specifying --host even though we are cross compiling - But this may
change in the future, so we better pass the correct values.

Passing extra jimtcl configure flags is unfortunately not forseen in the
usb_modeswitch Makefile, so add a small patch making this possible.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
(cherry picked from commit 29cccab4476bc58b2acb50887a67e310045efb24)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>